+++ source/adios2/toolkit/sst/cp/cp_internal.h
@@ -260,9 +260,9 @@ struct _CP_DP_PairInfo
 struct FFSFormatBlock
 {
     char *FormatServerRep;
-    int FormatServerRepLen;
+    size_t FormatServerRepLen;
     char *FormatIDRep;
-    int FormatIDRepLen;
+    size_t FormatIDRepLen;
     struct FFSFormatBlock *Next;
 };
